halcmd: allow net command to create anonymous nets
Anonymous nets are named 'n$0', 'n$1', and so forth.

This usage is not ambiguous, because until now it was an error
if the first argument of net had the same name as a signal.

However, any hal file that explicitly uses signals with names starting
'n$' will need to be rewritten to use a different prefix before it
can use anonymous nets.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>