configure: don't require linux/version.h
This has been a legacy/unsupported header file for a long time, and
according to memleak is finally gone in kernels somewhere in the 3.x
days.  We have long looked at several alternative filenames to actually
find the magic constants, but for some reason required that the
version.h header exist at a certain spot even though we didn't use it
except as a last resort.