machinetalk/nanopb: detect protobuf subtree falling out of sync
Since nanopb.proto is now part of machinetalk-protobuf, but is imported
as a git subtree into the machinekit repo, it might happen that an update
of the nanopb subtree brings in a newer nanopb.proto file from upstream.

This diff command checks against that happening, and aborts the build if
they fall out of sync.