gladevcp -tooledit, flush tool file to disk
hopefully fix an occational bug that would make linuxcnc
tell you that the tool was missing from file when it was obviously there.
It seems maybe the file was not written to disk before linuxcnc was
told to read it. why it would be _blank_ I don't know.
I could cause this reliable when testing wihout linuxcnc, probably
because of the big pause as it tried to talk to non-existant linuxcnc..

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>