docs: update stepper quickstart equation
The "source" for the equation was correct in the asciidoc .txt file, but
the image containing the rendered equation was out of date.  The latex
equation images are not automatically generated/updated, so i did it
by hand.

Thanks to Joe Hildreth for the bug report.

For next time, these are the steps i did:

* I uncommented the relevant latex part of the asciidoc .txt file.

* I ran 'make doc' to generate the html file with the latex source in
  a pass-through block.

* I ran html-latex-images on the html file to generate the updated
  .png files.

* I overwrote the .png in git with the regenerated one from the script
  and this is the commit.