gladevcp -fix blinking in GLADE editor problem with led widget
The first time you added an LED the introspection would fail
becaue blink_rate was set to None - a value outside GObjects int variable.
Now the rate is set to 0. If a user selects a rate between 1 and 99 it is set
to 100 ms. setting it to 0 turns off blinking.
The nice checkbutton in the editor for selecting blinking unfortunately
does not send a signal to LED.py to deselect blinking, only to select it.
so now one must set the rate to zero to stop blinking.
Another oddity of GLADE ....