Fix Boost detection for newer Debian derivatives
The Boost library is now multiarched, so the Boost libraries may live
in e.g. /usr/lib/x86_64-linux-gnu.  The GNU autoconf-archive
`ax_boost_base.m4` tests don't account for this as of 13 Sep 2014,
[git rev. 3d67989][1].

This commit looks for `dpkg-architecture`, and uses the
`DEB_HOST_MULTIARCH` variable to generate another subdirectory
candidate for search.

[1]: http://git.savannah.gnu.org/gitweb/?p=autoconf-archive.git;a=blob;f=m4/ax_boost_base.m4;h=d7c9d0d3986e4dd4556dad84809d0f0fabdbac6b;hb=HEAD