tests: fix a transient failure in the halui-jogging test
This test verifies that only the selected joint jogs.  But it doesn't
wait for the joint to stop jogging at the end of the test, so when it
goes on to the next joint it might see the tail end of the previous
joint's jog and fail.

This commit makes it wait for the joint to stop moving before going on
to test the next joint.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>