remove EMC_WAIT_NONE from shcom, it's broken
Before this commit it was possible for programs using shcom to set
emcWaitType to EMC_WAIT_NONE, which meant that no wait at all would
happen after a command was put on the Command NML channel.

The programs using shcom are:
    * linuxcncrsh
    * the Tcl linuxcnc library
    * linuxcnclcd
    * the schedrmt tool

This wait mode is bug-prone and means messages from these programs to
Task might get overwritten and lost.

This "none" wait type was never explicitly requested by any in-tree code
that I could find, but it's possible that out-of-tree code could use it.

This commit removes EMC_WAIT_NONE entirely, to ensure that it does not
get used.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>