lui: use a unique-ish serial number
If multiple lui instances are running, or if a process has both a lui
connection and an old-style "use NML directly" connection to linuxcnc,
then the Task might not notice a new message if it has the same serial
number as the previous message.

This commit attempts to spread out the lui serial numbers to make the
risk of collission smaller (but not 0).

And NML complains for serial numbers > 9,999,999, so multiply by 100 here.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>