lui: introduce, use get_command_channel_nml
We introduced get_status_nml so that UIs relying on an implicit update
of status after a lui api call would see it.  This fixes the opposite
problem: now a non-lui command would not cause an update of the lui
status buffer.  By using the same command channel, the problem is
eliminated.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>