build-depend on inkscape
We use dblatex in our docs build system, and it uses inkscape to convert
svg images to pdf for inclusion in the pdf docs.

Unfortunately the version of dblatex in Debian Wheezy (and probably other
distros) is missing the dependency on inkscape, so our pbuilder builds
of the pdf docs fail.  The resulting pdf documents are made with a text
string with the name of the svg file, instead of the image contents of
the svg file.

I opened Debian Bug #766945 to track this:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=766945

This commit adds to linuxcnc an explicit build-dependency on inkscape,
even though we dont "manually" use it during the build.  This explicit
build-dependency should be removed when we can expect a correct version
of dblatex.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>