task: reindent for consistency
This commit only changes whitespace, to get adjacent lines to line up
on the same column (given 8-character tabs).

Some nearby lines use tabs instead of spaces so the diff looks off,
but in an editor with 8-char tabs it looks right, and i didn't want to
replace all the tabs.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>