docs: minor reorg in the html index
This moves the Stepper Diagnostics from the Getting Started section to
the Integrator Manual section.

This makes the Getting Started section of the html index have nearly
the same contents as the Getting Started pdf document.

The only difference now is that the pdf document has some additional
copyright info, which is at the bottom of the html index rather than in
the Getting Started section (for now)...

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>