latency-histogram: clean up on ^C
Catch SIGINT (^C on the keyboard) and call the cleanup function to stop
realtime and unload kernel modules (instead of just exiting and leaving
things set up).

Since Tcl doesn't support signal handling natively we need to use Extended
Tcl, so depend on that...

Thanks to Peter Wallace for the bug report.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>