comp: don't overrun the names[] array
Before this commit, if a comp-generated component was given the max
number of names= arguments, it would happily walk past the end of the
names[] array until it came to a NULL.

In kernels since 2.6, module_param_array() sets the number of entries
actually provided, we should probably use that.  Needs SIM support
though.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>