canon: Fix for helical acceleration calculation
Solve the issue of slow arcs and helices on machines with lower
acceleration along the Z axis than other axes.

Previously, the maximum acceleration in the helical direction was
incorrectly calculated using the total path "time^2" and the normal axis
acceleration. Now, we use only the displacement / time along the normal
axis. This means that arcs with small movement along the normal axis
(i.e. simple arcs and gradual helices) will not be artificially limited
this way.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>