Fix metric arc tolerence to same value as imperial
Also add an ini file field and allow user variance of arc tolerances
within preset limits
Default tolerance set to figure 2 * 0.001 * sqrt(2) for inch,
and 2 * 0.01 * sqrt(2) for mm.

This would mean that any valid arc where the endpoints and/or
centerpoint got rounded or truncated to 0.001 inch or 0.01 mm
precision would be accepted.

This behavior is the root of the problem with some CAM generated code.

Minimum tolerance capped at 1 micron and equivilent imperial.

Signed-off-by: Mick Grant<arceye@mgware.co.uk>