hal: gantry.comp: Modify latching behavior
This restores the previous movement behavior, where joint offsets are
updated only when moving towards the home switches.  Adding latching
behavior when backing off the switches requires both the assertion and
release of the switch to be mechanically accurate

The gantry component is inteded to be used with search_vel and latch_vel
that have the same sign, and expects the assertion of the home signal to
be the accurate mechancial event.

Signed-off-by: Charles Steinkuehler <charles@steinkuehler.net>