tests: fix a spurious failure of the tlo test
Part of the tlo test runs a deliberately incorrect g-code and checks
for the expected error message from LinuxCNC.

It's possible that LinuxCNC generates an unrelated error messages
before we run the bad g-code.  Specifically, we sometimes get this one
in the buildbot:

    Unexpected realtime delay on task 0
    This Message will only display once per session.
    Run the Latency Test and resolve before continuing.

This unexpected error message would confuse the test and it would fail.

This commit tries to reduce the likelihood of a spurious test failure for
this reason by draining the error queue before running the bad g-code.
There's still a possibility that the unexpected error will be generated
after we drain the error queue but before we run the bad g-code, but
this does make things significantly better.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>