halcmd: manage prompt better
Only display a prompt if run in interactive mode, ie if the input is
a terminal.  As always, the prompt is "halcmd: " in normal mode and "%"
in script mode.

If the input is not a terminal, do not display a prompt.  This is new
behavior, prompted (heh) by ugly output (lots of "%" lines") from stepconf
running halcmd.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>