interp: Added a method to restore state on abort
The state tag changes show the "current" state according to motion.
However, an early abort will cause the displayed modes to suddenly
switch to the actual interpreter state. This could be confusing to a
user, so we call a restore method on abort that pulls the latest state
tag from motion and unpacks it into a usable state.