src/configure: remove the check for tclx
TclX is used by latency-histogram; without tclx installed it won't run.
TclX is used at runtime, not during the build, so the debian/control
Depends on tclx but does not Build-Depend on it.  So it's not guaranteed
to be installed, and src/configure should not count on it.

(In 2.7 and later, configure's --disable-check-runtime-deps flag *does*
let us check for tclx in normal user-initiated builds.)

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>