motion: fix for zero-length line errors
This patch adds a specific return code for zero-length lines in motion.
This error is has a positive actual value, so it is not treated as a
deal-breaking error. This is because a zero length line in principle is
something we're free to ignore at the motion level.

This specifically solves an issue Sam was seeing, where a motion segment
very close to the cutoff size (1e-7) was being treated as zero length.
The cause of this error was that homing was not returning the machine's
actual position to exactly zero (though it was within 1e-8).
Unfortunately, this difference was just enough to make a short segment a
bit too short by the standards of motion, so it was treated as zero
length, throwing an error.

This is a band-aid fix for now. The proper way would be to fix the
slight position disconnect between motion and interp.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>