axis gui: rename joint_mode to teleop_mode
This commit contains a variable & function rename only, no change in
behavior.

The internal variable "joint_mode" is True when the machine is in
Teleop mode, and False when the machine is in Free (aka Joint) mode.
This is confusing.

Similarly with the functions named set_joint_mode() and
toggle_joint_mode(), which also do the opposite of what you'd expect.

This commit renames everything so that the logic makes sense.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>