tp: Improved handling of low-queue state
The problem: Optimization falters when the queue only has a few segments
in it. Specifically, we can't be sure of the total segment length until
there is another segment after it (due to blending). The approach in the
past was very conservative, and basically aborted optimization entirely
if the queue had less than 3 segments in it.

This commit addresses this problem somewhat with 2 major changes:
1) Allow blending on segments already in progress, provided that a few
conditions are true (not doing parabolic blending already, less than
1/2 of the segment is complete).
2) Since a segment will never have more than 1/2 of it's length consumed
in a blend, the worst-case final velocity of the previous segment is the
maximum velocity that we can reach and still stop at the halfway point.

With these fixes, we don't have to fall back to parabolic blending at
the beginning and end of programs (at least not as frequently), so
execution is much more consistent. The trade-off with this approach is
that the last segment before a queue-buster or the end of a program has
slightly poorer blending performance.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>