tp: fix for "climbing" behavior in rigid tapping
It turns out we were always setting the tp goalPos to be the endpoint of
the most recently added segment. Unfortunately, this assumption breaks
on rigid tapping, where the "end point" is somewhere past the start
point (but motion stops early once the spindle isn't synced). This was a
subtle side-effect of a very early refactor of the trajectory planner
functions. The fix is a bit of a wart, but at least rigid tapping
behaves as expected now.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>