tp: polish and tweaks for low queue fix
Improved the test for whether we can safely blend a segment already in
progress. The worst case is that 1/2 of the segment is consumed by a
a very small blend (imagine a 179 deg. turn). This means that (in this
worst case) we need to be able to come essentially to a stop by the
halfway point of a segment. If the segment is already in progress, then
this safe cutoff is at 25% of the original segment (assuming a
triangular velocity profile). While we could optimize this a bit more,
the majority of cases of the segment already starting are only a few
timesteps, so we very rarely approach this limit except in the case of
very short segments, and even then the savings would not be great.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>