hallib/instance: new named HAL object 'instance'
An instance is basically a named blob of memory in the HAL shared memory segment,
and always owned by a component (like pins, params and functs so far).

An instance has an integer ID, which can serve as an owner_id (formerly
component id) of a pin, param, or funct.

vtables are comp-level objects only and hence unaffected.
rings are owned by hal_lib as they transcend the component and instance
lifecycle, and hence unaffected too.

A HAL instance is described by hal_inst_t.