tp: overhaul spiral fit computation to use more numerically stable quadratic formula
In the spiral arc fit, near-circular spirals have poorly conditioned
quadratic coefficients in the arc-length fit. The "standard" quadratic
formula becomes ill-conditioned  when the "A" coefficient is small. This
commit uses an alternate form of the quadratic formula that is robust
against small "A" terms, leading eliminating the need for the fit
tolerance used previously.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>