examples/userfuncts: demo for the userfunct API
userfuncts are a similar to normal functs, but cannot be addf'd to a thread.
However, they can be called from userland via halcmd, or the C or Cython RTAPI API's.

userfuncts are the basic feature on which the HAL instantiation API rests.
hal_lib is now a component and exports internal userfuncts (newinst and delinst)
to support the instantiation API.