axis: Use a preferred form of "switch" (closes: SF#411)
The Tcl 'switch' command has several forms.  The one formerly used in
update_state is *not* a preferred format, because the compiler can't
statically tell whether $::task_state is an option (if it starts with a
"-") or the string to switch on (if it doesn't).

The format used in the rest of our Tcl code, with all the body-pattern
pairs enclosed as a single Tcl list is not appropriate here because in
this case the patterns are the result of expanding $-variables.

That leaves the "explicit end-of-options" form, in which "--" is
inserted before the value to be matched against a pattern.  When this
form is used, it also happens to fix SF#411 on my Debian Jessie system
when using an affected version of tcl, such as tcl8.6=8.6.2+dfsg-2.

Since the leading theory for the cause of SF#411 is an obscure bug in
Tcl, changing how the Tcl compiler is able to reason about the proc--by
taking it from a non-preferred code path to a preferred one--is a
reasonable workaround.

It also means a traceback will be avoided in the abnormal case that
$::task_state holds a negative integer.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>