tp: relax pause requirement for reverse run
Pausing was a lazy way to ensure that the machine absolutely was not
moving when doing a direction switch. However, a properly-written
"is-moving" function can do this same job without needing an explicit
pause from the user. One critical side-effect of this is that we can
give the appearance of smooth transitions between forward and reverse
motion with adaptive feed.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>