configure: clean up cache files
I noticed that incorrect version information was shown in AXIS
after this sequence:
    checkout master
    autogen
    checkout 2.7
    autogen

This information, as well as other stuff, could end up cached in
autom4te.cache.  autogen.sh always intended to remove this cache but
it had the wrong filename; and rm -f doesn't signal an error when
the target doesn't exist.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>