debian/control: Temp fix for libzmq[34] problem
When ZeroMQ4 was released, the library version wasn't updated to `4`
until about 4.0.4. Perhaps because of this confusion, the Debian
package is still called `libzmq3` in Jessie despite this change now
being present.

The Dovetail Automata Wheezy package repository shipped a `libzmq4`
package, anticipating the library version bump would also come with a
package name bump.

This interim fix allows Machinekit to build against Jessie or the
Dovetail Automata Wheezy packages until the latter is brought into
sync, after which it can be removed.