halrun: depend on halcmd/haltcl to clean up realtime
There's some kind of bug or limitation in the handling of
suspend by a grandchild process of the shell.  This can be
observed e.g., by
  $ bash -c 'bash ; :'
  $ suspend # (this is the interactive grand child bash)
  [the terminal is hung]
With halrun, the problem manifests slightly differently:
  $ bash -c 'halcmd ; :'
  halcmd: ^Z
  $ # at this point halcmd is consuming 100% CPU
  $ fg
  [the terminal is hung]
The problem seems to be something to do with readline, because
  $ bash -c 'vim ; :'
doesn't seem to be affected and may ^Z - fg cycles may be
conducted without incident.  The problem is reported from
time to time, but the readline people don't seem to have a
definite solution to it.  e.g.,
https://lists.gnu.org/archive/html/bug-readline/2015-03/msg00000.html

I investigated a variety of systems, including some farily old
ones -- RedHat 9 (ca 2003), FreeBSD 6.2 (ca 2006) -- and they
all reproduce the "bash -c 'bash ; :'" version of the problem,
which I believe is a similar underlying problem from readline.

I think that possibly vim works differently due to this:
    The fact is that vim doesn't use the SIGTSTP mechanism at
    all, choosing instead to interpret ^Z as regular input.
    You can observe this by dumping the tty settings for a
    terminal running vim; you will find that isig is disabled.
[http://www.linusakesson.net/programming/vimsuspend/]

Anyway, all these problems arise when halcmd is a *grandchild*
of the terminal's interactive shell.  By introducing and using
the new "-r" (is-halrun) flag, halrun can *exec* the interactive
portion so that it is a child, not a grandchild, of the shell.
For whatever nonsense reason, this either fixes the problem
or makes it drop from hitting 100% of the time to hitting way
under 10% of the time.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>