rtapi: accommodate systems without sys/io.h
.. such as aarch64.  On these systems, there is no
suitable definition of inb/outb to be had, as there is no I/O
memory space distinct from the regular address space.

Implement these functions such that inX returns zero,
outX does nothing.

(In retrospect, it's surprising that 32-bit arm has
sys/io.h, inb/outb and the rest, but apparently for certain early
ARM systems such as the NetWinder, inb/outb were mapped onto a
certain range of memory addresses known as the "footbridge"; that
lives on in glibc to this day despite the fact that all the netwinders
are probably in landfills by now)

Signed-off-by: Jeff Epler <jepler@unpythonic.net>