tp: fix for tangent improvement
Stupid oversight, the optimizer was always clipping final velocity to
the kink velocity, even when we used the blend arc instead of
approximate tangent.

Now, the TP decides either to use a blend arc if it's faster, and clear the now-invalid kink
velocity, or fall back to approximate tangency (and use the
kink_velocity).

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>