stepconf -add import-mach file
This adds John T's mach conversion file after being modified
to be added to stepconf.
I think there are some small issues with this, some of them
I have partially addressed.
It seems to make assumptions about whether data is int or float.
It's not surprising as stepconf willl record data either way, depending
on the value at the time.
It disn't detect if the mach config had a A axis or was a lathe config.
This should be fixed now.
Some A axis data was not recorded - I fixxed some/all?
There was some math conversion problem such as 8.e+003 to 8000
that should be fixed now.
I am not familar with Mach so cannot test very thoughly.

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>