uspace: ensure that the thread-specific key is initialized
failing to do so caused rtapi_self *outside* a realtime thread
to occasionally return a nonnegative value, while this is
defined as needing to return a negative failure code.

As a consequence, the message
    hm2_eth: ERROR: used llio->read in realtime task
would occasionally be (spuriously) shown at the startup if hm2_eth was
loaded before a component that created a thread, such as threads or
motion.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>