docs: don't force image width for step type diagram
Forcing the width makes part of the image fall off the PDF page, making
it invisible.

This commit removes the width forcing so asciidoc will make it as big as
it needs to be to fit on the pdf page, though at the expense of making the
html image too wide on narrow screens (but at least that has scroll bars).

Thanks to Bernt Hustad Hembre for the bug report.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>