UIs: increase task receive timeout to 5.0 seconds
Task sometimes goes out to lunch for seconds and seconds, usually when
calling rename() as part of saving the interp parameter file.

Some UIs had short (order of 1 or 2 seconds) timeouts for Task
acknowledging commands they sent, which sometimes lead to test failures.

Notably different is shcom (used by linuxcncrsh and a few others),
which uses infinity as its task acknowledgement timeout.

This commit increases the timeout for all UIs to 5.0 seconds, which
should make the UIs work more reliably (but not perfectly, Task still
isn't realtime, and it might be distracted for >5.0 seconds at times).

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>