hy-vfd: fix bug with partial reads from VFD
"p_msg" points to the part of the packet we just read, "msg" points to
the beginning of the packet buffer.

We want to read the "packet length" field out of the header at a specific
offset from the *beginning* of the packet, not from wherever it is we
happened to do the most recent read.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>