debian/rules.in: make soft link for nc_files
so that one can start a config in
  /usr/share/doc/linuxcnc/examples/sample-configs/sim/
directly without copying files (see Notes below).

It was reported on irc (tom_itx) that a config would not start when linuxcnc was
invoked at the command line when using a path to an ini file like:
  $ linuxcnc /usr/share/doc/linuxcnc/examples/sample-configs/sim/name_of_config.ini

The config worked correctly when using the config picker which copies files
to the user directory named ~/linuxcnc/configs

This patch adds a required link to make the above example work.
Ref: configs/maintainer.txt for related information

Notes:
 1) The /usr/share/doc/linuxcnc/examples/sample-configs
    files are owned by root so writing and/or updating files (*.var for example)
    is not possible for non-root user but maybe they should start up anyway
 2) configs that attempt to create files in a root-owned ini-containing directory will
    not succeed (currently: probing and ngcgui examples, maybe others)

Signed-off-by: Dewey Garrett <dgarrett@panix.com>