gaxis -name some widgets so the theme can see them
estop and power on buttons are named so a modified theme
 can visually modifiy the buttons.
For whatever reason naming the gtktoggletoolbutton is not
enough so we use get_child() to name the togglebutton it
represents.
naming the widget for theming is not the same thing as
naming for id. By default Gscreen uses the widget id as the
widget's theming name.
(The id is the name given in the GLADE editor)

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>