Overhaul adaptive feed handling for reverse run
One difficulty of using negative adaptive feed to trigger reverse run is
that the actual scale factor has to be non-negative. With a few
additional checks, this commit handles the translation from a bilateral
input (-1.0 to 1.0) to the unilateral feed scale and direction switch
(forward or reverse).

What happens now is that a negative adaptive feed causes the TP to try
to reverse direction. If this is successful, or we're already in
reverse, then the scale factor is just the absolute value of the
adaptive feed pin. Otherwise, the scale factor is set to zero, so the TP
decelerates to a stop (allowing a direction change). A similar situation
happens for the reverse to forward transition

EDIT: fix merge conflicts 12 Dec 2018 <phillc54>

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>