tests: add another loadrt test
In 2.6 sim and 2.7 uspace, a failed loadrt of a component causes later
loadrt calls of the same component to fail with the error message
"already exists".

This is due to a bug in how sim_rtapi_app/uspace_rtapi_app do tracking
of loaded modules, fixed in the next commit.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>