Motion: set the "In Position" emcmot status flag when aborting
Before this commit, aborting an in-progress move would leave the "In
Position" flag in the Status buffer set to False, which is misleading.

The old behavior would lead to Axis blocking for 10 seconds in certain
situations:

    * start Axis, turn machine on, home
    * run the "splash screen" gcode
    * in the middle of a long move, hit E-stop
    * try to reload the file, or open a new file

Result: Axis hangs for 10 seconds.

The problem is that Axis sends Task Mode commands (switching between MDI
and Auto) before requesting a file open, and waits for receipt from Task
by checking the "status" field of the Status NML buffer.

Task sets the status field to RCS_EXEC while the machine is in motion.
Because Motion forgot to set the inpos field to True on abort, Task
would never go from RCS_EXEC to RCS_DONE, and Axis would have to wait
for its timeout.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>