tp: Remove dependency on binormal vector for spherical arc
Since arcs in more than 3 axes will not have a valid binormal vector, we
can't use the cross product to create a tangent vector to a blend arc.
The only time this happens currently is after the blend arc is created,
to verify that it's actually tangent to the other segments.

This implementation the derivative of the SLERP formula to calculate a
tangent vector. It's slightly more expensive than the cross product w/
binormal since it costs two cosine evaluations.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>