Implement G52 offsets
This implementation of `G52` is intended to be compatible with Fanuc,
including sharing offset registers with `G92`.

LinuxCNC rs274ngc persists `G92` offsets across machine startup, M02
and M30, but Fanuc does not.  Because users familiar with Fanuc may
expect the offset to be volatile, and be surprised (perhaps badly)
when the offset persists, this patch adds a `.ini` file option to
disable it.

Other implementation details:

- `G52` shares registers with `G92`, `#5210` (flag) and `#5211`
  through `#5219` (9 axis offsets)
- `G52` offsets are applied in the same way as `G92`, after workpiece
  coordinate offsets and rotation.
- `G52` and `G92`/`G92.n` may be freely mixed, and in the light of the
  shared registers, interactions behave as expected.
  - `G52 X0 Y0` zeros the offsets, but does not reset the `#5210`
    flag.  This is not expected to cause unexpected behavior; the flag
    apparently exists only to implement `G92.2` and `G92.3`.
  - `G52` and `G92` share DRO readouts and preview offset display.
- `G92` offset persistence may be disabled by setting
  `PERSISTENT_G92_OFFSET = false` in the `[RS274NGC]` section of the
  `.ini` file.  By default, `G92` is persistent.

Signed-off-by: John Morris <john@zultron.com>