task: correct axes based on axis mask
This number is eventually exposed in the stat buffer, and many UIs
have used it to choose how many axes to display.

The more modern axis mask is preferable, because it allows UIs to correctly
show the right axes when some are missing.  But for better operation of
old UIs (as well as code in task that may incorrectly rely on the value
of Axes), correct the setting so that it includes all axes specified
by the axis mask.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>