build-depend on w3c-linkchecker
This commit causes the HTML links to be validated every time the HTML
docs are built, using the W3C "checklink" program from the
w3c-linkchecker package.

This validation was optional before, because Lucid doesn't have the
w3c-linkchecker package, so we couldn't build-depend on it.  Now that we
don't support Lucid any more, we can finally add this build dependency.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>