Really ugly hacks to add UVW displacements to arc length calculation.
It's ugly mostly because it's the first cut. To make 6-axis blending
work (with UVW), path length has to be expressed as the total length
along the 6-dimensional trajectory. Before, the path length was only
computed in XYZ.

Note that this doesn't change the physical motion of the machine. All
motions in XYZUVW will happen just like they did before. The only
difference is how the path is represented internally. Ex:

G20 G91 G1 X1 Y1 Z1 U1 V1 W1 F60

This is a move of +1in for all linear axes.
The old way:

arc length = sqrt(1^2 + 1^2 + 1^2) = sqrt(3) ~= 1.732

The new way:

arc length = sqrt(1^2 + 1^2 + 1^2 + 1^2 + 1^2 + 1^2) = sqrt(6) ~= 2.45

The path velocity will be proportionally higher as well. To maintain a
feed of 60 ipm for XYZ motion, the TP's internal "velocity" will
actually be 1.414x higher.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>