features_preview: sim configs firsttime startup
sim configs need at least one .ngc file in myfeatures/lib/
so that linuxcnc can find a real directory in the SUBROUTINE_PATH on
the initial run with 'features' embedded.  Formerly, the
directory was created by features.py on the initial run (but
too late to be detected by linuxcnc).

also update .gitnore

Signed-off-by: Dewey Garrett <dgarrett@panix.com>