features_preview: dialog for system file updates
If an updated system file becomes available (due to a LinuxCNC
version upgrade for example), handle the situation where the user has
a modified local copy.  Use a dialog to offer to copy in the updated
system file(s) or keep user's locally modified file(s).

In the dialog, support yes,no for acceptance of each updated
system file and support refresh or cancel to accept or reject
_all_ newly available files.

If a locally modified file is kept, touch it to prevent a reissue
of the dialog on subsequent startups.

Note: a new system file (e.g., a file which has not existed
      previously) is silently copied to the user's directory

Also: remove unneeded share/features/scripts/
      (scripts/ dir is local use only)

Signed-off-by: Dewey Garrett <dgarrett@panix.com>