handle building with lxrt.so (RTAI 4.1 and newer)
Earlier versions of RTAI (at least the 3.8.1 and 3.9 versions we shipped
on Lucid, Wheezy, and Precise) inline the lxrt code in the header file.

RTAI 4.1 and later, including 5.0-test1 that's currently packaged for
Jessie, do not inline lxrt in the header, instead provides the lxrt
functionality as a library.

This means that it's not enough for our code to include the lxrt header
any more, we now need to link in the lxrt library.

This commit uses the standard rtai-config script to learn the CFLAGS and
LDFLAGS (which included LDLIBS) needed to build userspace stuff with lxrt,
and uses them when building liblinuxcnchal, which our only library that
directly interfaces with the RTAI userspace.

This is required on RTAI 4.1 and later, and seems benign on older
supported versions of RTAI:

Lucid (RTAI 3.8.1):
    lxrt-cflags: -I. -I/usr/realtime-2.6.32-122-rtai/include -O2 -I/usr/src/linux-headers-2.6.32-122-rtai/include -Wall -Wstrict-prototypes -pipe
    lxrt-ldflags: -L/usr/realtime-2.6.32-122-rtai/lib -lpthread

Wheezy/Precise (RTAI ~3.9):
    lxrt-cflags: -I. -I/usr/realtime-3.4-9-rtai-686-pae/include -O2 -I/usr/src/linux-headers-3.4-9-rtai-686-pae/include -I/usr/src/linux-headers-3.4-9-rtai-686-pae/include2 -Wall -Wstrict-prototypes -pipe
    lxrt-ldflags: -L/usr/realtime-3.4-9-rtai-686-pae/lib -lpthread

Jessie (RTAI 5.0-test1):
    lxrt-cflags: -I. -I/usr/include/rtai -I/usr/src/linux-headers-3.16.0-9-rtai-686-pae/include -Wall -Wstrict-prototypes -pipe
    lxrt-ldflags: -L/usr/lib -llxrt -lpthread

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>